Nope..
Former-commit-id: 65d00e07a8e84d96fc8ee2c3dbf96cbafb81352d

+++ run_camera_pose_estimation.py
@@ -147,7 +147,7 @@ def main():
     print(args)
 
     #Float or Double?
-    tensor_type = torch.double
+    tensor_type = torch.single
 
 
     #Load datasets
